+++ component/inc/board_link.h
@@ -52,7 +52,7 @@ i2c_addr_t component_id_to_i2c_addr(uint32_t component_id);
  * This function utilizes the simple_i2c_peripheral library to
  * send a packet to the AP and wait for the message to be received
 */
-void send_packet_and_ack(uint8_t len, uint8_t* packet);
+void send_packet_and_ack(uint16_t len, uint8_t* packet);
 
 /**
  * @brief Wait for a new message from AP and process the message
