+++ component/src/board_link.c
@@ -47,7 +47,21 @@ i2c_addr_t component_id_to_i2c_addr(uint32_t component_id) {
  * This function utilizes the simple_i2c_peripheral library to
  * send a packet to the AP and wait for the message to be received
 */
-void send_packet_and_ack(uint8_t len, uint8_t* packet) {
+void send_packet_and_ack(uint16_t len, uint8_t* packet) {
+    
+    // int i = 0;
+    // while (len > MAX_I2C_MESSAGE_LEN-1) {
+    //     I2C_REGS[TRANSMIT_LEN][0] = MAX_I2C_MESSAGE_LEN-1;
+    //     memcpy((void*)I2C_REGS[TRANSMIT], (void*)packet + i, MAX_I2C_MESSAGE_LEN-1);
+    //     I2C_REGS[TRANSMIT_DONE][0] = false;
+    //     // Wait for ack from AP
+    //     while(!I2C_REGS[TRANSMIT_DONE][0]);
+    //     I2C_REGS[RECEIVE_DONE][0] = false;
+
+    //     len -= MAX_I2C_MESSAGE_LEN-1;
+    //     i += MAX_I2C_MESSAGE_LEN-1; 
+    // }
+
     I2C_REGS[TRANSMIT_LEN][0] = len;
     memcpy((void*)I2C_REGS[TRANSMIT], (void*)packet, len);
     I2C_REGS[TRANSMIT_DONE][0] = false;
@@ -55,6 +69,7 @@ void send_packet_and_ack(uint8_t len, uint8_t* packet) {
     // Wait for ack from AP
     while(!I2C_REGS[TRANSMIT_DONE][0]);
     I2C_REGS[RECEIVE_DONE][0] = false;
+
 }
